[refactor] add newsEmbed to reduce repetitiveness in news files
