+++ src/commands/news/edit.ts
@@ -2,22 +2,19 @@ import { getNews, updateNews } from "database/news";
 import { getSetting } from "database/settings";
 import {
   ContainerBuilder,
-  MediaGalleryBuilder,
-  MediaGalleryItemBuilder,
   SlashCommandSubcommandBuilder,
   TextDisplayBuilder,
   type ChatInputCommandInteraction,
   type InteractionResponse,
   type Message,
-  type Role,
   type TextChannel,
 } from "discord.js";
 import { errorEmbed } from "embeds/errorEmbed";
+import { newsEmbed } from "embeds/newsEmbed";
 import { colorize, Sokolors } from "utils/colorize";
-import { mention } from "utils/mention";
 import { modalSubmit } from "utils/modalSubmit";
 import { newsModal } from "utils/newsModal";
-import { safeChannel, safeMember, safeRole } from "utils/safeThings";
+import { safeChannel, safeMember } from "utils/safeThings";
 import { sendChannelNews } from "utils/sendChannelNews";
 
 export const data = new SlashCommandSubcommandBuilder()
@@ -64,56 +61,19 @@ export async function run(
   const modalInteraction = await modalSubmit(interaction);
   if (!modalInteraction) return;
 
-  const role = (await getSetting(guild.id, "news", "role")) as string;
-  const roleToSend: Role | null = role ? await safeRole(guild, role) : null;
-
   const title = modalInteraction.fields.getTextInputValue("title");
   const body = modalInteraction.fields.getTextInputValue("body");
-  const avatar = news.authorPFP;
-  const image = news.imageURL;
   const editedContainer = new ContainerBuilder()
     .addTextDisplayComponents(new TextDisplayBuilder().setContent("## News post edited."))
     .setAccentColor(await colorize({ hue: Sokolors.Green }));
 
   if (!(await getSetting(guild.id, "news", "edit_original_message"))) {
-    await sendChannelNews(
-      guild,
-      interaction,
-      {
-        title,
-        body,
-        author: news.author,
-        authorPFP: avatar,
-        id,
-      },
-      true,
-    );
+    await sendChannelNews(guild, interaction, { title, body, author: news.author, id }, true);
     return await modalInteraction.reply({
       components: [editedContainer],
       flags: ["Ephemeral", "IsComponentsV2"],
     });
   }
-
-  const container = new ContainerBuilder()
-    .addTextDisplayComponents(
-      new TextDisplayBuilder().setContent(
-        `**Posted by ${news.author}${roleToSend ? ` for ${roleToSend}` : ""}**`,
-      ),
-      new TextDisplayBuilder().setContent(`## ${title}`),
-      new TextDisplayBuilder().setContent(body),
-    )
-    .setAccentColor(await colorize({ hue: Sokolors.Blue }));
-
-  if (image)
-    container.addMediaGalleryComponents(
-      new MediaGalleryBuilder().addItems(new MediaGalleryItemBuilder().setURL(image)),
-    );
-
-  container.addTextDisplayComponents(
-    new TextDisplayBuilder().setContent(
-      `-# Edited news post from ${guild.name} • ID: ${news.id} • ${mention(news.updatedAt?.valueOf() ?? news.createdAt.valueOf(), "DEFAULT_TIMESTAMP")}`,
-    ),
-  );
 
   const channel = (await safeChannel(
     guild,
@@ -121,7 +81,10 @@ export async function run(
   )) as TextChannel;
 
   await Promise.all([
-    channel.messages.edit(news.messageID, { components: [container], flags: "IsComponentsV2" }),
+    channel.messages.edit(news.messageID, {
+      components: [await newsEmbed(guild, { title, body, author: news.author, id }, true)],
+      flags: "IsComponentsV2",
+    }),
     updateNews(guild.id, id, title, body),
     modalInteraction.reply({
       components: [editedContainer],
